requirements.txt
requests
azure-common==1.1.28
azure-core==1.30.1
azure-storage-blob==12.19.1
python-dotenv